+++ src/renderer/src/components/ModelLibrary/providers/index.ts
@@ -16,12 +16,14 @@ export * from './ZaiPlanParameters';
 
 // Advanced Settings
 export * from './OpenAiAdvancedSettings';
+export * from './OpenAiCompatibleAdvancedSettings';
 export * from './GeminiAdvancedSettings';
 export * from './VertexAiAdvancedSettings';
 export * from './RequestyAdvancedSettings';
 
 // Model Overrides
 export * from './OpenAiModelOverrides';
+export * from './OpenAiCompatibleModelOverrides';
 export * from './GeminiModelOverrides';
 export * from './VertexAiModelOverrides';
 export * from './RequestyModelOverrides';